Fix Windows warnings in application code
Update the QFileInfo hash overload to match Qt 6 hash signatures so
64-bit MSVC no longer narrows the size_t hash value through uint.

Initialize MSVC math constants before Qt math headers in OCCT-enabled
translation units, mark the Windows native window fallback parameter as
intentionally unused, and move the navbar splitter to the Qt 6 mouse
position API.

+++ src/partcache.h
@@ -37,7 +37,7 @@ private slots:
     void onDirectoryChange(const QString &path);
 };
 
-inline uint qHash(const QFileInfo &key, uint seed)
+inline size_t qHash(const QFileInfo &key, size_t seed)
 {
     return qHash(key.absoluteFilePath(), seed);
 }
